0-read_textfile 6

+++ 0x15-file_io/main.h
@@ -2,5 +2,5 @@
 #define MAIN_H
 
 ssize_t read_textfile(const char *, size_t);
-void *err_malloc(int);
+void *err_malloc(size_t);
 #endif
